fix: use numbers instead of string for codec
The codec used to be defined as string. Now you can use the numbers as
defined per multicodec constants. The codec is stored in a field
called `code`.

Ideally we would store it in `codec`, but for backwards compatibility
reasons that one will still return the name of the codec as a string.

New code should use `code` instead of `codec` to check for the codec
a CID is using.