fix: create new CID from old CID
When creating a CID from another CID instance that was created by and old version of this module (< 0.7) the `multibaseName` property would become `undefined` and cause an error to be thrown when calling `toString`. This PR ensures an appropriate `multibaseName` is picked in the edge case that this happens.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>